Extracted `html-to-plaintext` shared lib into package
