+++ ghost/core/package.json
@@ -72,6 +72,7 @@
     "@tryghost/errors": "1.2.14",
     "@tryghost/express-dynamic-redirects": "0.0.0",
     "@tryghost/helpers": "1.1.71",
+    "@tryghost/html-to-plaintext": "0.0.0",
     "@tryghost/image-transform": "1.1.0",
     "@tryghost/job-manager": "0.0.0",
     "@tryghost/kg-card-factory": "3.1.3",
@@ -143,7 +144,6 @@
     "glob": "8.0.3",
     "got": "9.6.0",
     "gscan": "4.32.0",
-    "html-to-text": "8.2.0",
     "human-number": "2.0.0",
     "image-size": "1.0.2",
     "intl": "1.2.5",
