+++ ghost/html-to-plaintext/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./lib/html-to-plaintext');
