+++ ghost/html-to-plaintext/lib/html-to-plaintext.js
