+++ ghost/html-to-plaintext/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@tryghost/html-to-plaintext",
+  "private": true,
+  "version": "0.0.0",
+  "author": "Ghost Foundation",
+  "license": "MIT",
+  "main": "index.js",
+  "scripts": {
+    "dev": "echo \"Implement me!\"",
+    "test": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura mocha './test/**/*.test.js'",
+    "lint": "eslint . --ext .js --cache"
+  },
+  "files": [
+    "index.js",
+    "lib"
+  ],
+  "devDependencies": {
+    "c8": "7.12.0",
+    "mocha": "10.0.0",
+    "should": "13.2.3",
+    "sinon": "14.0.0"
+  },
+  "dependencies": {
+    "html-to-text": "8.2.0",
+    "lodash": "4.17.21"
+  }
+}
